tests/wpt/tests/css/css-grid/alignment/grid-align-baseline-003.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/css-align-3/#baseline-rules">
<link rel="help" href="https://drafts.csswg.org/css-grid-1/#alignment">
<style>
#target {
  display: grid;
  grid-auto-flow: column;
  grid-template-rows: 120px;
  align-items: last baseline;
  width: 200px;
  border: solid 3px;
  position: relative;
  line-height: 0;
}
span {
  display: inline-block;
  width: 1em;
  height: 1em;
  outline: solid cornflowerblue 3px;
  outline-offset: -3px;
}
#target > div {
  flex: none;
}
#target > :nth-child(1) {
  background: lime;
  margin-bottom: 20px;
  padding-bottom: 20px;
  font-size: 20px;
}
#target > :nth-child(2) {
  background: hotpink;
  font-size: 30px;
}
#target > :nth-child(3) {
  background: papayawhip;
  font-size: 10px;
  writing-mode: vertical-lr;
}
</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<body onload="checkLayout('#target > div')">

<div id="target">
  <div data-offset-y="40"><span></span><br><span></span></div>
  <div data-offset-y="20"><span></span><br><span></span></div>
  <div data-offset-y="70"><span></span><br><span></span></div>
</div>
