LayoutTests/imported/w3c/web-platform-tests/css/css-grid/alignment/grid-align-baseline-003-expected.txt





FAIL #target > div 1 assert_equals:
<div data-offset-y="40"><span></span><br><span></span></div>
offsetTop expected 40 but got 10
FAIL #target > div 2 assert_equals:
<div data-offset-y="20"><span></span><br><span></span></div>
offsetTop expected 20 but got 0
FAIL #target > div 3 assert_equals:
<div data-offset-y="70"><span></span><br><span></span></div>
offsetTop expected 70 but got 20
